¡Lots of BugFixes!

+++ src/file_io.h
@@ -1,9 +1,10 @@
 #ifndef COMPILER_FILE_IO_H
 #define COMPILER_FILE_IO_H
 
+#include <stddef.h>
 #include <stdio.h>
 
-long file_size(FILE *file);
+size_t file_size(FILE *file);
 char *file_contents(char *path);
 
 #endif /* COMPILER_FILE_IO_H */
